Fixed next-js build error, Type error: Route "src/app/api/auth/[...nextauth]/route.tsx" does not match the required types of a Next.js Route.   "authOptions" is not a valid Route export field.
