+++ src/lib/session.ts
@@ -1,5 +1,5 @@
 import { getServerSession } from "next-auth/next";
-import {authOptions} from "../app/api/auth/[...nextauth]/route";
+import { authOptions } from "./authOptions";
 
 
 export async function getCurrentUser() {
